feature: optional until date time
This feature allows a user to optionally add a third parameter to the
script which captures an "until" date and passes it to the git cli. This
will allow users to run reports for quarters or time lines in the past.
If the parameter is left empty it will default to the current date time.