Add team totals and delta
This change to cloc-author-date shows a Team Total and Delta for all
the changes. This will help with faster metrics as team metrics used
to need to be added up manually.

The final report was updated to be cleaner by removing blank results.
Comments were also added to the core script to denote what needs to
be commented out in the event of running multiple reports in a timely
manner if ./cloc-tmp contains a recent git pull from a first run